requirements.txt
boto3==1.4.7
botocore==1.7.36
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.11.6
django-extensions==1.9.1
django-storages==1.6.5
djangorestframework==3.7.3
docutils==0.14
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
jmespath==0.9.3
olefile==0.44
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
s3transfer==0.1.11
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
